Remove the HTML wrapper, add --language_out
This commit removes the HTML output wrapper. While it was useful for
generating a full HTML page with one command, it is easy to replicate
with external tools and not very flexible. In the future, it might make
sense to include a more useful template once we know common use
patterns.

The --language_out parameter can be used to target older versions of
JavaScript, if it is needed.